IA no longer shows All possibles paths only the current path and fixed an out of range exception with the pathfinding when there is no way to the objective

+++ Assets/Scripts/CaminoCompleto.cs
@@ -19,7 +19,7 @@ public class CaminoCompleto
         }
     }
 
-    public void ShowPath(int duration) {
+    public void ShowPath(float duration) {
         foreach (Nodo actual in caminoNodo) {
             Debug.DrawLine(actual.transform.position + Vector3.up, actual.direccion.transform.position + Vector3.up, Color.cyan, duration);
         }
